requirements.txt
streamlit>=1.55,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
duckdb>=1.0,<2.0
scikit-learn>=1.5,<2.0
google-genai>=1.0.0
python-dateutil>=2.9,<3.0